--- pkg/convert/pprof/bench/.gitignore
@@ -1,1 +1,0 @@
-*
